non_auth_user

+++ sis_book_cicle/book/views.py
@@ -7,12 +7,12 @@ from datetime import date
 import datetime
 
 
-def pagic(list_posts, posts_of_page=9):
+def pagic(list_posts, posts_of_page=9): #Функция для пагинации страничек
     '''Функция паджинатор'''
     return Paginator(list_posts, posts_of_page)
 
 
-def Litic(author):
+def Litic(author): #Функция для работы с Литкоинами
     '''Функция для монет'''
     if Litcoin.objects.filter(author=author.id):
         return Litcoin.objects.filter(author=author.id)[0]
@@ -20,7 +20,7 @@ def Litic(author):
         return 0
 
 
-def index(request):
+def index(request): #Функция главной страницы сайта
     '''Функция отвечающая за главную страницу'''
     books_list = Book.objects.filter(availability=True)
     title = 'Каталог'
@@ -28,34 +28,36 @@ def index(request):
     page_obj = pagic(books_list).get_page(page_number)
     coin = Litic(request.user)
 
-    rn = Rentsbook.objects.filter(arendator=request.user)
+    if request.user.is_authenticated is True:
 
-    ob = []
-    for r in rn:
-        ob.append(r)
+        rn = Rentsbook.objects.filter(arendator=request.user)
 
-    if len(ob) != 0:
-        rn = rn[0]
-        today = datetime.date.today()
-        time_to_td = rn.end_rent - today
+        ob = []
+        for r in rn:
+            ob.append(r)
 
-        resp = (f'У вас осталось {time_to_td.days} '
+        if len(ob) != 0:
+            rn = rn[0]
+            today = datetime.date.today()
+            time_to_td = rn.end_rent - today
+
+            resp = (f'У вас осталось {time_to_td.days} '
                 f'дней до конца аренды {rn.book.name}!')
 
-        if time_to_td.days == 1:
-            book = rn.book
-            rn.delete()
-            book.delete()
+            if time_to_td.days == 1:
+                book = rn.book
+                rn.delete()
+                book.delete()
 
-        context = {
-            'title': title,
-            'books': books_list,
-            'page_obj': page_obj,
-            'coin': coin,
-            'resp': resp
-        }
+            context = {
+                'title': title,
+                'books': books_list,
+                'page_obj': page_obj,
+                'coin': coin,
+                'resp': resp
+            }
 
-        return render(request, 'books/test_index.html', context)
+            return render(request, 'books/test_index.html', context)
 
     context = {
         'title': title,
@@ -67,50 +69,59 @@ def index(request):
     return render(request, 'books/test_index.html', context)
 
 
-def book_detail(request, book_id):
+def book_detail(request, book_id): #Функция для детаьлной странички книги
     '''Функция отвечающая за страницу книги'''
     book = Book.objects.filter(id=book_id)
     book = book[0]
     coin = Litic(request.user)
-    rent = Rentsbook.objects.filter(user=request.user, book=book)
-    rent_arendator = Rentsbook.objects.filter(
-        arendator=request.user,
-        book=book
-    )
 
-    published_posts = []
+    if request.user.is_authenticated is True:
 
-    published_posts_are = []
+        rent = Rentsbook.objects.filter(user=request.user, book=book)
+        rent_arendator = Rentsbook.objects.filter(
+            arendator=request.user,
+            book=book
+        )
 
-    for post in rent:
-        published_posts.append(post)
+        published_posts = []
 
-    for post in rent_arendator:
-        published_posts_are.append(post)
+        published_posts_are = []
 
-    if len(published_posts) != 0:
+        for post in rent:
+            published_posts.append(post)
+
+        for post in rent_arendator:
+            published_posts_are.append(post)
+
+        if len(published_posts) != 0:
+            context = {
+                'rent': rent[0],
+                'coin': coin,
+            }
+            return render(request, 'books/rent_book.html', context)
+
+        if len(published_posts_are) != 0:
+            context = {
+                'rent': rent_arendator[0],
+                'coin': coin,
+            }
+            return render(request, 'books/rent_book.html', context)
+
         context = {
-            'rent': rent[0],
+            'book': book,
             'coin': coin,
         }
-        return render(request, 'books/rent_book.html', context)
-
-    if len(published_posts_are) != 0:
-        context = {
-            'rent': rent_arendator[0],
-            'coin': coin,
-        }
-        return render(request, 'books/rent_book.html', context)
-
+        return render(request, 'books/test_book.html', context)
+    
     context = {
         'book': book,
-        'coin': coin,
     }
+    
     return render(request, 'books/test_book.html', context)
 
 
 @login_required
-def book_create(request):
+def book_create(request): #Функция для добавления книги на сайт
     '''Функция отвечающая за создание поста'''
     coin = Litic(request.user)
     form = BookForm(request.POST or None,
@@ -133,7 +144,7 @@ def book_create(request):
 
 
 @login_required
-def profile(request, username):
+def profile(request, username): #Фунция для профиля
     '''Функция отвечающая за страницу профиля'''
     author = User.objects.get(username=username)
     book_list = Book.objects.filter(author=author)
@@ -148,7 +159,7 @@ def profile(request, username):
 
 
 @login_required
-def book_rent(request, book_id):
+def book_rent(request, book_id): #Функция для запроса аренды книги
     '''Функция отвечающая за аренду книги'''
     book = get_object_or_404(Book, id=book_id)
     title = 'Каталог'
@@ -186,7 +197,7 @@ def book_rent(request, book_id):
 
 
 @login_required
-def book_rent_agree(request, username):
+def book_rent_agree(request, username): #Функция для соглашения на запрос аренды
     '''Функция отвечающая за аренду книги'''
     usersnames = User.objects.filter(username=username)[0]
     rent_user = Rentsbook.objects.filter(user=usersnames, in_rent=0)
@@ -203,7 +214,7 @@ def book_rent_agree(request, username):
 
 
 @login_required
-def user_rent(request, username):
+def user_rent(request, username): #Функция отвечающая за страницу арендованных человеком книг
     '''Функция отвечающая за страницу арендованных человеком книг'''
     user = User.objects.get(username=username)
     rents = Rentsbook.objects.filter(arendator=user, in_rent=1)
@@ -221,7 +232,7 @@ def user_rent(request, username):
     return render(request, 'books/profile_rent.html', context)
 
 
-def book_one_or_two_agree(request, book_id):
+def book_one_or_two_agree(request, book_id): #Функция для подтверждения передачи с двух сторон
     '''Функция отвечающая за подтверждение передачи'''
     book = get_object_or_404(Book, id=book_id)
     cost = book.coins
